pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brooks-trading-coach"
version = "0.1.0"
description = "Brooks Price Action Trading Coach - Advisory system for discretionary day traders"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Trading Coach Team"}
]
keywords = ["trading", "price-action", "brooks", "journal", "analytics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",
    "sqlalchemy>=2.0.25",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "yfinance>=0.2.36",
    "mplfinance>=0.12.10b0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
llm = [
    "openai>=1.12.0",
]
web = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "bcrypt>=4.1.0",
    "PyJWT>=2.8.0",
    "pydantic[email]>=2.5.0",
]
rag = [
    "PyMuPDF>=1.24.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.0",
]
databento = [
    "databento>=0.30.0",
]
robinhood = [
    "robin_stocks>=3.0.0",
]

[project.scripts]
brooks = "app.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    "_deps",
    "_pip_cache",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
